requirements.txt
# base server
Django==3.2
djangorestframework==3.12.4
djangorestframework-filters==1.0.0.dev2
psycopg2>=2.9.1,<2.9.2
Pillow>=8.3.2,<8.3.3
requests
pandas

flake8>=3.6.0<3.7.0
ipdb==0.13.7